requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
autopep8==2.1.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
diskcache==5.6.3
distro==1.9.0
docker==7.0.0
docopt==0.6.2
docutils==0.21.1
FLAML==2.1.2
Flask==3.0.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
importlib_metadata==7.1.0
itsdangerous==2.1.2
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.0
Jinja2==3.1.3
keyring==25.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.2.0
nh3==0.2.17
numpy==1.26.4
openai==1.16.2
packaging==24.0
pkginfo==1.10.0
pyautogen==0.2.23
pycodestyle==2.11.1
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
python-dotenv==1.0.1
PyYAML==6.0.1
readme_renderer==43.0
regex==2023.12.25
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.1
setuptools==69.3.0
sniffio==1.3.1
termcolor==2.4.0
tiktoken==0.6.0
tqdm==4.66.2
twine==5.0.0
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.2
zipp==3.18.1
